DD-123: Changed the color scheme to incorporate a dark background with lighter text. Checking the accessibility tab in Firefox's developer console shows that the only contrast issues that we have are with the sample Tableau chart's color choices, so that'll get fixed with Mohit's awesome chart.
DD-114: Incorporated a new sidebar feature and moved the h1 element of the dashboard to within tableau.component.html to make it work better overall (fullscreen attribute would make the h1 element disappear otherwise). Clicking on the "Options" button will bring up the accessibility options sidebar that allows for future options like light mode vs. dark mode to be toggled (or to go to different public Tableau books if we decide to go with the separate workbook option).

No JIRA: I liked the color scheme that Will used for the primary and secondary colors, so I moved their definition to the root of styles.css so the whole app can use them. styles.css would need to be loaded first to ensure that these colors are loaded properly, however.

No JIRA: Also did CTRL+ALT+L to format the files with consistent formatting (hence why a lot of files are in green now).

Known bug: Because overflow definitions for the mat-sidenav-content element gets changed whenever the sidebar is opened, Tableau flickers everytime since it redraws the chart. A fix for this is required to improve UX (easy ticket for Sprint #3 since UX is a ticket there).